DIGG-58: Refactoring to single file search page
