--- pages/concepts/index.ts
@@ -1,3 +1,0 @@
-import { SearchTermsPage } from "@/components/content/Search/SearchPage/SearchTermsPage";
-
-export default SearchTermsPage;
