+++ pages/concepts/index.tsx
@@ -1,0 +1,8 @@
+import { NextPage } from "next";
+import { SearchPageEntryScape } from "@/components/content/Search/SearchPage/SearchPageEntryScape";
+
+const ConceptsPage: NextPage = () => (
+  <SearchPageEntryScape searchType="concepts" />
+);
+
+export default ConceptsPage;
