--- pages/datasets/index.ts
@@ -1,3 +1,0 @@
-import { SearchDataPage } from "@/components/content/Search/SearchPage/SearchDataPage";
-
-export default SearchDataPage;
