+++ pages/datasets/index.tsx
@@ -1,0 +1,8 @@
+import { NextPage } from "next";
+import { SearchPageEntryScape } from "@/components/content/Search/SearchPage/SearchPageEntryScape";
+
+const DataSetPage: NextPage = () => (
+  <SearchPageEntryScape searchType="datasets" />
+);
+
+export default DataSetPage;
