--- pages/specifications/index.ts
@@ -1,3 +1,0 @@
-import { SearchSpecificationsPage } from "@/components/content/Search//SearchPage/SearchSpecificationsPage";
-
-export default SearchSpecificationsPage;
