+++ pages/specifications/index.tsx
@@ -1,0 +1,8 @@
+import { NextPage } from "next";
+import { SearchPageEntryScape } from "@/components/content/Search//SearchPage/SearchPageEntryScape";
+
+const SpecificationsPage: NextPage = () => (
+  <SearchPageEntryScape searchType="specifications" />
+);
+
+export default SpecificationsPage;
